10_inject_testContainer : chore : packages in lower cases, unit, acceptance, and integration tests correctly segregated, debug logger level, injecting testContainer when integration testing
