+++ src/main/resources/application.properties
@@ -1,1 +1,2 @@
 spring.data.mongodb.uri=mongodb://mongo-dakar:27017/mongo-dakar
+logging.level.root=DEBUG
