+++ src/main/resources/graphQLSchemas/journey.graphqls
