+++ src/test/java/com/dakar/dakar/unit/DakarApplicationUnitTests.java
@@ -1,0 +1,41 @@
+package com.dakar.dakar.unit;
+
+import com.dakar.dakar.models.Journey;
+import com.dakar.dakar.repositories.JourneyRepository;
+import com.dakar.dakar.services.JourneyService;
+import lombok.extern.slf4j.Slf4j;
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
+import org.mockito.junit.MockitoJUnitRunner;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import java.util.List;
+
+import static org.junit.Assert.assertNotNull;
+
+@Slf4j
+@RunWith(MockitoJUnitRunner.class)
+@SpringBootTest
+public class DakarApplicationUnitTests {
+
+    @InjectMocks
+    private JourneyService journeyService;
+
+    @Mock
+    private JourneyRepository journeyRepository;
+
+    @Before
+    public void setup() {
+        MockitoAnnotations.initMocks(this);
+    }
+
+    @Test
+    public void gotAllJourney() {
+        List<Journey> journeyList = journeyService.allJourney();
+        assertNotNull(journeyList);
+    }
+}
