+++ src/test/resources/application.properties
@@ -1,0 +1,2 @@
+spring.data.mongodb.uri=thisPropoertyWillBeModifiedAtRunTimeInTheSpringContext
+logging.level.root=DEBUG
